Fixed lots of warnings.
Rev: src/cpp.c:1.7
Rev: src/lex.c:1.38
Rev: src/main.c:1.34
Rev: src/modules/Image/encodings/gif.c:1.26
Rev: src/modules/Regexp/pike_regexp.c:1.10
Rev: src/modules/spider/spider.c:1.50
Rev: src/pike_types.c:1.28
Rev: src/signal_handler.c:1.21

+++ src/signal_handler.c
@@ -259,7 +259,7 @@ static RETSIGTYPE receive_signal(int signum)
 #endif
 #endif
 #endif
-    fprintf(stderr,"pid %d died with code %d\n",pid,status);
+    fprintf(stderr,"pid %d died with code %d\n",(int)pid,status);
 
     if(pid>0)
     {
